Add more guards.
Move some of it to the AbstractHandler so that they get reusable.
