+++ src/ContaoCommunityAlliance/DcGeneral/Contao/View/Contao2BackendView/ActionHandler/AbstractHandler.php
@@ -12,8 +12,11 @@
 
 namespace ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\ActionHandler;
 
+use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\Exception\EditOnlyModeException;
+use ContaoCommunityAlliance\DcGeneral\Data\ModelIdInterface;
 use ContaoCommunityAlliance\DcGeneral\Event\ActionEvent;
 use ContaoCommunityAlliance\DcGeneral\EnvironmentInterface;
+use ContaoCommunityAlliance\DcGeneral\Exception\DcGeneralRuntimeException;
 
 /**
  * Handler class for handling the show events.
@@ -61,6 +64,41 @@ abstract class AbstractHandler
     protected function getEnvironment()
     {
         return $this->getEvent()->getEnvironment();
+    }
+
+    /**
+     * Guard that the environment is prepared for models data definition.
+     *
+     * @param ModelIdInterface $modelId The model id.
+     *
+     * @throws DcGeneralRuntimeException If data provider name of modelId and definition does not match.
+     */
+    protected function guardValidEnvironment(ModelIdInterface $modelId)
+    {
+        if ($this->getEnvironment()->getDataDefinition()->getName() !== $modelId->getDataProviderName()) {
+            throw new DcGeneralRuntimeException(
+                sprintf(
+                    'Not able to perform action. Environment is not prepared for model "%s"',
+                    $modelId->getSerialized()
+                )
+            );
+        }
+    }
+
+    /**
+     * Guard that the data container is not in edit only mode.
+     *
+     * @param ModelIdInterface $modelId The model id.
+     *
+     * @return void
+     *
+     * @throws EditOnlyModeException If data container is in edit only mode.
+     */
+    protected function guardNotEditOnly(ModelIdInterface $modelId)
+    {
+        if ($this->getEnvironment()->getDataDefinition()->getBasicDefinition()->isEditOnlyMode()) {
+            throw new EditOnlyModeException($modelId->getDataProviderName());
+        }
     }
 
     /**
